+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
-* Update to 1.20.6
+* New functionality: Ignore stems when holding axe
+    * This option prevents you from destroying stems when holding an axe (Useful for manual pumpkin and melon farming)
+* Reworked hit mobs through non-solid block